backend/requirements.txt
fastapi
uvicorn
transformers
torch
torchaudio
pillow
python-multipart
ffmpeg-python
